# app/static/swagger.yaml
openapi: 3.0.0
info:
  title: "Mechanics Shop API"
  description: "A comprehensive REST API for managing automotive service operations including customers, service tickets, mechanics, and inventory"
  version: "1.0.0"
  contact:
    name: "API Support"
    email: "support@mechanicshop.com"
  license:
    name: "MIT"
    url: "https://opensource.org/licenses/MIT"
servers:
  - url: "http://127.0.0.1:5000"
    description: "Development server"
  - url: "https://api.mechanicshop.com"
    description: "Production server"
components:
  securitySchemes:
    customerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Customer JWT Token"
    mechanicAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Mechanic JWT Token"
  schemas:
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Error message describing what went wrong"
        details:
          type: object
          description: "Additional error details"
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operation completed successfully"

    # Customer Schemas
...
    MechanicResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        first_name:
          type: string
          example: "John"
        last_name:
          type: string
          example: "Doe"
        email:
          type: string
          format: email
          example: "john.doe@mechanicshop.com"

    # Service Ticket Schemas
    ServiceTicketCreate:
      type: object
      required:
        - customer_id
        - description
      properties:
        customer_id:
          type: integer
          example: 1
          description: "ID of the customer creating the ticket"
        description:
          type: string
          example: "Brake repair and oil change needed"
          description: "Description of the service required"
        status:
          type: string
          enum: [open, in_progress, completed, cancelled]
          default: "open"
          example: "open"
        mechanic_ids:
          type: array
          items:
            type: integer
          example: [1, 2]
          description: "Array of mechanic IDs to assign to this ticket"
...
    ServiceTicketResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        description:
          type: string
          example: "Brake repair and oil change"
        customer_id:
          type: integer
          example: 1
        status:
          type: string
          example: "open"
        mechanics:
          type: array
          items:
            $ref: '#/components/schemas/MechanicResponse'
        inventory:
          type: array
          items:
            $ref: '#/components/schemas/InventoryResponse'
    TicketMechanicsUpdate:
      type: object
      properties:
        add_ids:
          type: array
          items:
            type: integer
          example: [1, 2]
          description: "Mechanic IDs to add to the ticket"
        remove_ids:
          type: array
          items:
            type: integer
          example: [3]
          description: "Mechanic IDs to remove from the ticket"
    TicketPartAdd:
      type: object
      required:
        - part_id
      properties:
        part_id:
          type: integer
          example: 1
          description: "Inventory part ID to add to the ticket"

    # Inventory Schemas
...
    InventoryResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "Brake Pads"
        price:
          type: number
          format: float
          example: 49.99
        created_at:
          type: string
          format: date-time
          example: "2023-01-15T10:30:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2023-01-15T10:30:00Z"

    # Pagination and List Schemas
...
paths:
...
  /tickets:
    post:
      tags:
        - Service Tickets
      summary: "Create Service Ticket"
      description: "Create a new service ticket. Requires customer authentication."
      security:
        - customerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceTicketCreate'
            examples:
              newTicket:
                summary: "New Service Ticket"
                value:
                  customer_id: 1
                  description: "Brake repair and oil change needed"
                  status: "open"
                  mechanic_ids: [1, 2]
      responses:
        201:
          description: "Service ticket created successfully"
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Service ticket created successfully"
                  data:
                    $ref: '#/components/schemas/ServiceTicketResponse'
        400:
          description: "Bad Request - Missing required fields"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags:
        - Service Tickets
      summary: "Get All Service Tickets"
      description: "Retrieve all service tickets with caching and rate limiting. Requires authentication."
      security:
        - customerAuth: []
      responses:
        200:
          description: "Tickets retrieved successfully"
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ServiceTicketResponse'
                  count:
                    type: integer
                    example: 10
  /tickets/{id}:
    get:
      tags:
        - Service Tickets
      summary: "Get Service Ticket by ID"
      description: "Retrieve a specific service ticket by ID with caching. Requires authentication."
      security:
        - customerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
          description: "Service Ticket ID"
      responses:
        200:
          description: "Ticket found"
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/ServiceTicketResponse'
        404:
          description: "Ticket not found"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Service Tickets
      summary: "Delete Service Ticket"
      description: "Delete a service ticket. Requires mechanic authentication."
      security:
        - mechanicAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
          description: "Service Ticket ID"
      responses:
        200:
          description: "Ticket deleted successfully"
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Service ticket deleted successfully"
        404:
          description: "Ticket not found"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /tickets/{id}/edit:
    put:
      tags:
        - Service Tickets
      summary: "Edit Ticket Mechanics"
      description: "Add or remove mechanics from a service ticket"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
          description: "Service Ticket ID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TicketMechanicsUpdate'
            examples:
              updateMechanics:
                summary: "Update Ticket Mechanics"
                value:
                  add_ids: [1, 2]
                  remove_ids: [3]
      responses:
        200:
          description: "Ticket updated successfully"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        404:
          description: "Ticket not found"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /tickets/{id}/add-part:
    post:
      tags:
        - Service Tickets
      summary: "Add Part to Ticket"
      description: "Add an inventory part to a service ticket"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
          description: "Service Ticket ID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TicketPartAdd'
            examples:
              addPart:
                summary: "Add Part to Ticket"
                value:
                  part_id: 1
      responses:
        200:
          description: "Part added to ticket successfully"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        404:
          description: "Ticket or part not found"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ========== INVENTORY ENDPOINTS ==========
...